django-skeleton-cat/requirements.txt
Django>=2.2.5
django-allauth>=0.42.1
django-crispy-forms>=1.9.1
django-debug-toolbar>=2.2
django-simpleui>=4.0.2
# django-taggit>=1.3.0
# django-hreflang>=2.2
# django-rosetta>=0.9.2
# django-cors-headers==3.1.0
# django-filter==2.2.0
# awesome-slugify==1.6.5
djangorestframework>=3.9.1
django_taggit>=2.2.6
django-taggit-serializer>=0.1.7



